% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/species_data_functions.R
\name{add_estimated_sds}
\alias{add_estimated_sds}
\title{Estimate missing records for standard deviation based on mean of body size}
\usage{
add_estimated_sds(clean_size_data, sd_pars)
}
\arguments{
\item{clean_size_data}{dataframe of species' masses and standard deviations; as generated by \code{clean_sp_size_data}}

\item{sd_pars}{parameters as list of \verb{$slope}, \verb{$intercept}; as generated by \code{get_sd_parameters}}
}
\value{
a dataframe of species' \code{species_id} (which matches the AOU in the Breeding Bird Survey), \code{mass} mean body mass, \code{sd} standard deviation body mass, and a new column for \code{estimated_sd}, a \code{TRUE}/\code{FALSE} flag for whether the standard deviation has been estimated using the parameters provided in \code{sd_pars}.
}
\description{
Fill in missing records for the standard deviation of body mass for a species, based on its mean body size and the parameters estimated by the linear model fit by \code{get_sd_parameters}.
}
\keyword{internal}
