% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/community_data_fxns.R
\name{community_generate}
\alias{community_generate}
\title{Generate samples community-wide}
\usage{
community_generate(community_tables)
}
\arguments{
\item{community_tables}{result of \link{prepare_community}}
}
\value{
dataframe
}
\description{
Generate samples community-wide
}
\examples{
community_data <- new_hartford_raw \%>\%
dplyr::filter(year == 1994) \%>\%
filter_bbs_survey() \%>\%
prepare_community()

community_generate(community_data)
}
