% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary_functions.R
\name{community_summarize}
\alias{community_summarize}
\title{Compute summary statistics for a community}
\usage{
community_summarize(community, id_vars = NULL, species_vars = NULL)
}
\arguments{
\item{community}{result of \link{community_generate}}

\item{id_vars}{vector of column names to group by}

\item{species_vars}{vector of column names to summarize over}
}
\value{
summarized community data frame
}
\description{
Compute summary statistics for a community
}
\examples{
bbs_route <- new_hartford_raw \%>\% filter_bbs_survey()
community_data <- community_generate(bbs_route)
community_summarize(community_data)

}
