% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/demo_route_raw.R
\docType{data}
\name{demo_route_raw}
\alias{demo_route_raw}
\title{Raw data for a hypothetical Breeding Bird Survey route.}
\format{
A data frame with 1470 rows and 15 variables:
\describe{
\item{record_id}{inherited}
\item{routedataid}{inherited}
\item{countrynum}{inherited}
\item{statenum}{inherited}
\item{route}{inherited}
\item{rpid}{inherited}
\item{year}{inherited}
\item{aou}{inherited}
\item{count10}{inherited}
\item{count20}{inherited}
\item{count30}{inherited}
\item{count40}{inherited}
\item{count50}{inherited}
\item{stoptotal}{inherited}
\item{speciestotal}{inherited}
}
}
\usage{
demo_route_raw
}
\description{
For inclusion in package vignettes, function testing.
}
\references{
\itemize{
\item{Pardieck, K. L., Ziolkowski, D. J., Lutmerding, M., Aponte, V., & Hudson, M.-A. (2019). North American Breeding Bird Survey Dataset 1966—2018, version 2018.0. U.S. Geological Survey. https://doi.org/10.5066/P9HE8XYJ}
}
}
\keyword{datasets}
