% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulate_populations.R
\name{draw_population}
\alias{draw_population}
\title{Draw individuals to make a population}
\usage{
draw_population(
  species_mean = NULL,
  species_sd = NULL,
  species_abundance = NULL
)
}
\arguments{
\item{species_mean}{mean body size}

\item{species_sd}{standard deviation of body size}

\item{species_abundance}{number of individuals to draw}
}
\value{
vector of individuals' simulated body masses
}
\description{
Draw individuals to make a population
}
