% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/species_data_functions.R
\name{estimate_sd}
\alias{estimate_sd}
\title{Estimate a species' standard deviation of body mass based on its mean body mass}
\usage{
estimate_sd(sp_mean, pars = NULL)
}
\arguments{
\item{sp_mean}{mean body mass, in grams}

\item{pars}{list containing \verb{$slope} and \verb{$intercept}, generated by \code{\link[=get_sd_parameters]{get_sd_parameters()}}. If not provided, estimated by running \code{get_sd_parameters(raw_masses)}.}
}
\value{
the estimated standard deviation of body mass
}
\description{
Using the parameters estimated in \code{get_sd_parameters}, estimate the standard deviation of body mass for a species based only on its mean body mass.
}
\keyword{internal}
