% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/species_data_functions.R
\name{filter_bbs_survey}
\alias{filter_bbs_survey}
\title{Filter BBS survey data to remove non-target species}
\usage{
filter_bbs_survey(bbs_survey_data)
}
\arguments{
\item{bbs_survey_data}{data frame with columns for species and aou}
}
\value{
bbs_survey_data with nightbirds, waterbirds, non-targets removed
}
\description{
Code derived from MATSS, originally derived from Harris et al 2018
}
\examples{

filter_bbs_survey(new_hartford_raw)

}
