% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bbs_species_list_functions.R
\name{find_nontarget_species}
\alias{find_nontarget_species}
\title{Find nontarget species}
\usage{
find_nontarget_species(species_table)
}
\arguments{
\item{species_table}{The species list for the Breeding Bird Survey.}
}
\value{
dataframe of nontarget species AOUs and names
}
\description{
For transparency and testing, not for use.
}
\keyword{internal}
