% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulate_populations.R
\name{generate_individuals}
\alias{generate_individuals}
\title{Draw individuals to make a population.}
\usage{
generate_individuals(
  species_mean = NULL,
  species_sd = NULL,
  species_abundance = NULL
)
}
\arguments{
\item{species_mean}{mean body size}

\item{species_sd}{standard deviation of body size}

\item{species_abundance}{number of individuals to draw}
}
\value{
vector of individuals' simulated body masses
}
\description{
This is not a user-facing function; rather it is the random number generator under-the-hood for \link{simulate_population}.
}
