% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary_functions.R
\name{identify_richness_designator}
\alias{identify_richness_designator}
\title{Identify the column to use to compute species richness}
\usage{
identify_richness_designator(community)
}
\arguments{
\item{community}{result of \link{community_generate}}
}
\value{
\code{community} having identified the best-guess column for species richness
}
\description{
Identify the column to use to compute species richness
}
\keyword{internal}
