% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metabolic_rate.R
\name{individual_metabolic_rate}
\alias{individual_metabolic_rate}
\title{Estimate individual-level basal metabolic rate}
\usage{
individual_metabolic_rate(mass)
}
\arguments{
\item{mass}{mass in grams}
}
\value{
estimated basal metabolic rate using pars from Fristoe 2015
}
\description{
Estimate individual-level basal metabolic rate
}
\examples{
individual_metabolic_rate(10)
}
