% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bbs_species_list_functions.R
\name{is_unidentified}
\alias{is_unidentified}
\title{Check for unidentified species}
\usage{
is_unidentified(names)
}
\arguments{
\item{names}{species name}
}
\value{
logical
}
\description{
Derived from MATSS, originally derived from Harris et al 2018 (weecology/bbs-forecasting)
}
\keyword{internal}
