% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{known_species}
\alias{known_species}
\title{List of species known to \code{birdsize}}
\format{
A data frame with 443 rows and 6 variables:
\describe{
\item{AOU}{AOU used in Paradieck et al. (2019)}
\item{genus}{genus, from Paradieck et al. (2019)}
\item{species}{species, from Paradieck et al. (2019)}
}
}
\usage{
known_species
}
\description{
A table of the AOU (Pardieck et al 2019), genus, and species of the 443 species with species-specific data built in to \code{birdsize}.
}
\references{
\itemize{
\item{Pardieck, K.L., Ziolkowski Jr., D.J., Lutmerding, M., Aponte, V., and Hudson, M-A.R., 2019, North American Breeding Bird Survey Dataset 1966 - 2018 (ver. 2018.0): U.S. Geological Survey, Patuxent Wildlife Research Center, https://doi.org/10.5066/P9HE8XYJ.}
}
}
\keyword{export}
