% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulate_populations.R
\name{lookup_species_pars}
\alias{lookup_species_pars}
\title{Look up species mean and sd body size given species ID}
\usage{
lookup_species_pars(species_code = NULL, genus = NULL, species = NULL)
}
\arguments{
\item{species_code}{species_ID as specified in the Breeding Bird Survey}

\item{genus}{genus}

\item{species}{species}
}
\value{
data frame with columns species_id, genus, species, mean_mass, mean_sd, contains_estimates
}
\description{
Look up species mean and sd body size given species ID
}
