% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vignette_fxns.R
\name{niceOutput}
\alias{niceOutput}
\title{Format a data frame nicely for printing}
\usage{
niceOutput(data_table)
}
\arguments{
\item{data_table}{a data frame}
}
\value{
formatted datatable
}
\description{
Format a data frame nicely for printing
}
\examples{

niceOutput(raw_masses)

}
