% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary_functions.R
\name{pop_summarize}
\alias{pop_summarize}
\title{Summary statistics for a population}
\usage{
pop_summarize(population)
}
\arguments{
\item{population}{dataframe generated by \link{pop_generate}.}
}
\value{
dataframe of summary metrics for population
}
\description{
Summary statistics for a population
}
\examples{
a_population <- pop_generate(10, mean_size= 20)
pop_summarize(a_population)
}
