% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/community_data_fxns.R
\name{prepare_community}
\alias{prepare_community}
\title{Prepare community data for sim fxns}
\usage{
prepare_community(community_data_table, abundance_column_name = "speciestotal")
}
\arguments{
\item{community_data_table}{dataframe containing at least EITHER \code{aou} or \code{mean} and a column for species abundances}

\item{abundance_column_name}{character, the name of the column with species abundances. Defaults to "speciestotal"}
}
\value{
list of two dataframes: $ids_table contains all columns in community_data_table not passed to sim functions, and $sim_vars_table contains columns to pass to the sim functions.
}
\description{
Prepare community data for sim fxns
}
\examples{
community_data <- new_hartford_raw \%>\%
dplyr::filter(year == 1994) \%>\%
filter_bbs_survey()

prepare_community(community_data)
}
