% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/species_define.R
\docType{class}
\name{Species-class}
\alias{Species-class}
\alias{Species}
\title{An S4 class to hold species information}
\description{
An S4 class to hold species information
}
\section{Slots}{

\describe{
\item{\code{aou}}{The AOU}

\item{\code{genus}}{genus}

\item{\code{species}}{species}

\item{\code{mean_size}}{size}

\item{\code{sd_size}}{sd of size}

\item{\code{sd_method}}{known, or estimated}

\item{\code{id}}{if AOU is provided, AOU; otherwise an integer for keeping track of hypothetical species}
}}

