% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{unidentified_species}
\alias{unidentified_species}
\title{Table of AOUs corresponding to unidentified species.}
\format{
A data frame with 72 rows and 4 variables:
\describe{
\item{AOU}{AOU}
\item{english_common_name}{English common name}
\item{genus}{genus}
\item{species}{species}
}
}
\usage{
unidentified_species
}
\description{
This data table is used internally to filter out unidentified species based on their AOUs. It is derived from the species table available in Pardieck et al. (2019)
}
\references{
\itemize{
\item{Pardieck, K. L., Ziolkowski, D. J., Lutmerding, M., Aponte, V., & Hudson, M.-A. (2019). North American Breeding Bird Survey Dataset 1966—2018, version 2018.0. U.S. Geological Survey. https://doi.org/10.5066/P9HE8XYJ}
}
}
\keyword{internal}
