% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/species_data_functions.R
\name{add_estimated_sds}
\alias{add_estimated_sds}
\title{Add estimates for SDs}
\usage{
add_estimated_sds(clean_size_data, sd_pars)
}
\arguments{
\item{clean_size_data}{cleaned data}

\item{sd_pars}{parameters as list of slope, intercept}
}
\value{
clean_size_data with filled in sds + column to flag if an estimate was added
}
\description{
Add estimates for SDs
}
