% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/species_data_functions.R
\name{clean_sp_size_data}
\alias{clean_sp_size_data}
\title{Reconcile taxonomic updates between 2008 and 2019}
\usage{
clean_sp_size_data(raw_size_data)
}
\arguments{
\item{raw_size_data}{dataframe of species' mean and standard deviation body sizes; use the included \code{raw_masses} data table.}
}
\value{
dataframe of species' mean and standard deviation body sizes, with name mismatches resolved.
}
\description{
Six species have undergone name changes or other minor taxonomic rearrangements between 2008 and 2019, resulting in name mismatches between data from the Breeding Bird Survey (Paradieck et al. 2019) and the CRC Handbook (Dunning 2008). This function resolves those mismatches such that all species in the Breeding Bird Survey are associated with the appropriate size records from the CRC Handbook.
}
\references{
\itemize{
\item{Dunning, J. B. (2008). CRC handbook of avian body masses (2nd ed.). CRC Press.}
\item{Pardieck, K. L., Ziolkowski, D. J., Lutmerding, M., Aponte, V., & Hudson, M.-A. (2019). North American Breeding Bird Survey Dataset 1966—2018, version 2018.0. U.S. Geological Survey. https://doi.org/10.5066/P9HE8XYJ}
}
}
\keyword{internal}
