% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/species_data_functions.R
\name{clean_sp_size_data}
\alias{clean_sp_size_data}
\title{Clean raw size data}
\usage{
clean_sp_size_data(raw_size_data)
}
\arguments{
\item{raw_size_data}{raw size data}
}
\value{
cleaned data
}
\description{
Updates names of species whose names have changed since 2010.
}
