% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/community_data_fxns.R
\name{community_generate}
\alias{community_generate}
\title{Generate samples community-wide}
\usage{
community_generate(
  community_data_table,
  abundance_column_name = "speciestotal"
)
}
\arguments{
\item{community_data_table}{dataframe containing at least one of \code{aou} or \code{mean} and a column for species abundances}

\item{abundance_column_name}{character, the name of the column with species abundances. Defaults to "speciestotal"}
}
\value{
dataframe
}
\description{
Generate samples community-wide
}
\examples{

community_generate(new_hartford_clean)
}
