% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary_functions.R
\name{community_summarize}
\alias{community_summarize}
\title{Compute summary statistics for a community}
\usage{
community_summarize(
  community,
  level = c("year", "species", "species_and_year", "custom"),
  id_vars = NULL
)
}
\arguments{
\item{community}{result of \link{community_generate}}

\item{level}{one of: "year", "species", "species_and_year", "custom"}

\item{id_vars}{vector of column names to group by, if using "custom" level}
}
\value{
summarized community data frame
}
\description{
Compute summary statistics for a community
}
\examples{
bbs_route <- demo_route_raw \%>\% filter_bbs_survey()
community_data <- community_generate(bbs_route)
community_summarize(community_data)

}
