% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/species_data_functions.R
\name{estimate_sd}
\alias{estimate_sd}
\title{Estimate SD given mean and optional pars}
\usage{
estimate_sd(sp_mean, pars)
}
\arguments{
\item{sp_mean}{mean to estimate from}

\item{pars}{as list of slope, intercept}
}
\value{
estimated sd
}
\description{
Estimate SD given mean and optional pars
}
\keyword{internal}
