% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulate_community.R
\name{filter_bbs_survey}
\alias{filter_bbs_survey}
\title{Clean raw Breeding Bird Survey survey data}
\usage{
filter_bbs_survey(bbs_survey_data)
}
\arguments{
\item{bbs_survey_data}{data frame with columns for species and AOU}
}
\value{
bbs_survey_data with unidentified species, nightbirds, waterbirds, non-targets removed
}
\description{
The raw data for the Breeding Bird Survey includes unidentified species and some species that are not well-sampled by the BBS methods. This function filters a dataframe to remove those species.
}
\examples{

head(filter_bbs_survey(demo_route_raw))

}
