% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulate_community.R
\name{filter_bbs_survey}
\alias{filter_bbs_survey}
\title{Filter BBS survey data to remove non-target species}
\usage{
filter_bbs_survey(bbs_survey_data)
}
\arguments{
\item{bbs_survey_data}{data frame with columns for species and aou}
}
\value{
bbs_survey_data with unidentified species, nightbirds, waterbirds, non-targets removed
}
\description{
Filter BBS survey data to remove non-target species
}
\examples{

filter_bbs_survey(demo_route_raw) \%>\%
head()

}
