% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/species_data_functions.R
\name{get_sd_parameters}
\alias{get_sd_parameters}
\title{Get SD parameters from linear model}
\usage{
get_sd_parameters(raw_size_data)
}
\arguments{
\item{raw_size_data}{raw size data}
}
\value{
list of slope and intercept from lm
}
\description{
Get SD parameters from linear model
}
