% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/species_data_functions.R
\name{get_sp_mean_size}
\alias{get_sp_mean_size}
\title{Get species' mean size and sd}
\usage{
get_sp_mean_size(sd_dat)
}
\arguments{
\item{sd_dat}{Dat with estimated sds}
}
\value{
Summarized to species mean size and sd
}
\description{
Get species' mean size and sd
}
