% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ind_draw.R
\name{ind_draw}
\alias{ind_draw}
\title{Draw individuals to make a population.}
\usage{
ind_draw(
  species_mean = NA_real_,
  species_sd = NA_real_,
  species_abundance = NA_integer_
)
}
\arguments{
\item{species_mean}{mean body size}

\item{species_sd}{standard deviation of body size}

\item{species_abundance}{number of individuals to draw}
}
\value{
vector of individuals' simulated body masses
}
\description{
This is not a user-facing function; it is the random number generator under-the-hood for \link{pop_generate}.
}
\keyword{internal}
