% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metabolic_rate.R
\name{individual_metabolic_rate}
\alias{individual_metabolic_rate}
\title{Estimate individual-level BMR}
\usage{
individual_metabolic_rate(mass)
}
\arguments{
\item{mass}{mass in grams}
}
\value{
estimated basal metabolic rate
}
\description{
Given an individual's body mass (in grams), use allometric scaling (Fristoe
2015) to estimate basal metabolic rate.
}
\examples{
individual_metabolic_rate(10)
}
\references{
\itemize{ \item{Fristoe, T. S. (2015). Energy use by migrants and
residents in North American breeding bird communities. Global Ecology and
Biogeography, 24(4), 406–415. https://doi.org/10.1111/geb.12262} }
}
