% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulate_populations.R
\name{pop_generate}
\alias{pop_generate}
\title{Simulate body masses for a population}
\usage{
pop_generate(
  abundance = NA,
  aou = NA,
  genus = NA,
  species = NA,
  mean_size = NA,
  sd_size = NA,
  sim_species_id = 1
)
}
\arguments{
\item{abundance}{integer number of individuals to draw. \emph{Required}.}

\item{aou}{aou}

\item{genus}{genus}

\item{species}{species}

\item{mean_size}{numeric, mean body mass (in grams) for this species.}

\item{sd_size}{numeric, standard deviation of body mass for this species.}

\item{sim_species_id}{defaults AOU or 1}
}
\value{
a dataframe with \code{abundance} rows - one record per individual - and columns for species attributes.

Specifically:
\itemize{
\item \code{aou}: the AOU, if provided
\item \code{sim_species_id}: the \code{sim_species_id} if provided
\item \code{genus}: the genus associated with the AOU if provided, or the genus if provided
\item \code{species}: the species associated with the AOU if provided, or the species if provided
\item \code{individual_mass}: the simulated body mass (in grams) for this individual
\item \code{individual_bmr}: the simulated basal metabolic rate for this individual
\item \code{mean_size}: the mean body mass for this species (i.e. the parameter used for simulation)
\item \code{sd_size}: the standard deviation of body mass for this species (i.e. the parameter used for simulation)
\item \code{abundance}: the number of individuals simulated of this species (i.e. parameter used for simulation)
\item \code{sd_method}: the method for finding the standard deviation for body mass for this species
}
}
\description{
Draws body mass measurements for a population of birds (of all the same species) given the population size and either (1) the species AOU or (2) the mean and potentially standard deviation of body mass for that species.
}
\details{
\code{abundance} is required, as well as \emph{one of}: \code{aou}, \code{genus} and \code{species}, or \code{mean_size}.
}
\examples{

pop_generate(abundance = 5, aou = 2881)
pop_generate(abundance = 5, genus = "Selasphorus", species = "calliope")
pop_generate(abundance = 5, mean_size = 20, sd_size = 3)

}
