% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/species_define.R
\name{species_define}
\alias{species_define}
\title{Define a species}
\usage{
species_define(
  aou = NULL,
  genus = NULL,
  species = NULL,
  mean_size = NULL,
  sd_size = NULL,
  id = 1
)
}
\arguments{
\item{aou}{AOU}

\item{genus}{genus}

\item{species}{species}

\item{mean_size}{mean body size}

\item{sd_size}{sd of body size}

\item{id}{identifier; if using taxonomic info, defaults to AOU. If not, defaults to 1. Supplying other values can be useful for simulation models.}
}
\value{
object of class Species
}
\description{
Populates a \link{Species} object with taxonomic/identifying information and parameters for mean and standard deviation of body mass. Order of preference: AOU > genus + species > user provided mean and sd > user provided mean and estimated sd.
}
