% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/species_define.R
\name{species_define}
\alias{species_define}
\title{Define a species}
\usage{
species_define(
  aou = NA,
  scientific_name = NA,
  mean_size = NA,
  sd_size = NA,
  sim_species_id = 1
)
}
\arguments{
\item{aou}{AOU}

\item{scientific_name}{the species' scientific name, as "Genus species"}

\item{mean_size}{mean body size}

\item{sd_size}{sd of body size}

\item{sim_species_id}{identifier; if using taxonomic info, defaults to AOU. If not, defaults to 1. Supplying other values can be useful for simulation models.}
}
\value{
list with species parameter information
}
\description{
Creates a list with taxonomic/identifying information and parameters for mean and standard deviation of body mass.
}
\details{
The identifying information used depends on which parameters are provided, with the following order of preference: AOU > scientific name > user provided mean and sd > user provided mean and estimated sd.
}
\examples{
species_define(aou = 2881)
species_define(scientific_name = "Perdix perdix")
species_define(mean_size = 400, sd_size = 30)
species_define(mean_size = 400)
}
