% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/species_define.R
\name{species_lookup}
\alias{species_lookup}
\title{Look up species mean and sd body size given species ID}
\usage{
species_lookup(species_code = NULL, genus = NULL, species = NULL)
}
\arguments{
\item{species_code}{AOU species code as specified in the Breeding Bird Survey}

\item{genus}{genus}

\item{species}{species}
}
\value{
data frame with columns aou, genus, species, mean_mass, mean_sd, contains_estimates
}
\description{
Look up species mean and sd body size given species ID
}
