% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/species_define.R
\name{species_lookup}
\alias{species_lookup}
\title{Species lookup}
\usage{
species_lookup(aou = NULL, genus = NULL, species = NULL)
}
\arguments{
\item{aou}{AOU species code as specified in the Breeding Bird Survey}

\item{genus}{genus}

\item{species}{species}
}
\value{
data frame with columns aou, genus, species, mean_mass, mean_sd, contains_estimates
}
\description{
Given either AOU or genus and species, looks up a species' taxonomic information and mean and standard deviation of body size in \link{sd_table}.
}
\examples{
species_lookup(aou = 2881)
species_lookup(genus = "Selasphorus", species = "calliope")
}
